Fix slack bot with document set (#1588)
Also includes a Persona db layer refactor
